+++ e2e/tests/plugins.e2e.js
@@ -259,7 +259,7 @@ describe('plugins', () => {
       cy.openSoundCloudModal().addSoundCloud();
       cy.shrinkPlugin();
       cy.waitForVideoToLoad();
-      cy.getEditor()
+      cy.focusEditor()
         .type('{uparrow}')
         .type('Will this fix the flakiness?');
       cy.eyesCheckWindow(this.test.title);