+++ packages/wrapper/web/src/ViewerModal.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import getImagesData from 'wix-rich-content-fullscreen/dist/lib/getImagesData';
+import getImagesData from 'wix-rich-content-fullscreen/dist/lib/getImagesData.cjs.js';
 import Fullscreen from 'wix-rich-content-fullscreen';
 import './viewermodal.global.css';